+++ lib.esm/hds/index.d.ts
@@ -1,4 +1,5 @@
 import { HD } from './hd';
+import { AlgorandHD } from './algorand';
 import { BIP32HD } from './bip32';
 import { BIP44HD } from './bip44';
 import { BIP49HD } from './bip49';
@@ -16,5 +17,5 @@ export declare class HDS {
     static getHDClass(name: string): typeof HD;
     static isHD(name: string): boolean;
 }
-export { HD, BIP32HD, BIP44HD, BIP49HD, BIP84HD, BIP86HD, BIP141HD, CardanoHD, ElectrumV1HD, ElectrumV2HD, MoneroHD };
+export { HD, AlgorandHD, BIP32HD, BIP44HD, BIP49HD, BIP84HD, BIP86HD, BIP141HD, CardanoHD, ElectrumV1HD, ElectrumV2HD, MoneroHD };
 //# sourceMappingURL=index.d.ts.map