+++ lib.esm/hds/monero.js
@@ -22,7 +22,7 @@ export class MoneroHD extends HD {
     constructor(options = {
         minor: 1, major: 0
     }) {
-        super(options);
+        super({ ecc: SLIP10Ed25519MoneroECC, ...options });
         const network = ensureTypeMatch(options.network, Network, { otherTypes: ['string'] });
         const networkType = network.isValid ? network.value.NAME : options.network;
         if (!Monero.NETWORKS.isNetwork(networkType)) {